[ADIÇÃO[ Grid dos alimentos funcionando normalmente
